Move test directory and add snapshot test
