+++ __tests__/bookshop/srv/admin-service.cds
