+++ package.json
@@ -14,7 +14,8 @@
   ],
   "scripts": {
     "lint": "npx eslint .",
-    "test": "jest --ci --collectCoverage"
+    "test": "jest --ci --collectCoverage",
+    "cds:version": "cds v -i"
   },
   "devDependencies": {
     "eslint": "^8",
@@ -23,7 +24,7 @@
   "peerDependencies": {
     "@cap-js/asyncapi": "^1.0.0",
     "@cap-js/openapi": "^1.0.2",
-    "@sap/cds": "7.9.4",
+    "@sap/cds": "^8.1.1",
     "@sap/cds-compiler": "5.0.6"
   }
 }
